+++ diesel_compile_tests/tests/compile-fail/upsert_cannot_be_nested.rs
@@ -15,26 +15,54 @@ table! {
 #[table_name="users"]
 struct NewUser(#[column_name(name)] &'static str);
 
+#[allow(deprecated)]
 fn main() {
     use self::users::dsl::*;
     let connection = PgConnection::establish("postgres://localhost").unwrap();
 
-    insert_into(users).values(&NewUser("Sean").on_conflict_do_nothing().on_conflict_do_nothing()).execute(&connection);
-    //~^ ERROR E0277
-    insert_into(users).values(&NewUser("Sean").on_conflict(id, do_nothing()).on_conflict_do_nothing()).execute(&connection);
-    //~^ ERROR E0277
-    insert_into(users).values(&NewUser("Sean").on_conflict_do_nothing().on_conflict(id, do_nothing())).execute(&connection);
-    //~^ ERROR E0277
-    insert_into(users).values(&NewUser("Sean").on_conflict(id, do_nothing()).on_conflict(id, do_nothing())).execute(&connection);
-    //~^ ERROR E0277
-    insert_into(users).values(&vec![NewUser("Sean").on_conflict_do_nothing()]).execute(&connection);
-    //~^ ERROR E0277
-    insert_into(users).values(&vec![&NewUser("Sean").on_conflict_do_nothing()]).execute(&connection);
-    //~^ ERROR no method named `execute`
-    //~| ERROR E0277
-    insert_into(users).values(&vec![&NewUser("Sean").on_conflict(id, do_nothing())]).execute(&connection);
-    //~^ ERROR no method named `execute`
-    //~| ERROR E0277
-    insert_into(users).values(&(name.eq("Sean").on_conflict_do_nothing(),)).execute(&connection);
-    //~^ ERROR E0277
+    insert_into(users)
+        .values(&NewUser("Sean")
+                .on_conflict_do_nothing()
+                .on_conflict_do_nothing());
+                //~^ ERROR no method named `on_conflict_do_nothing` found
+    insert_into(users)
+        .values(&NewUser("Sean").on_conflict_do_nothing())
+        .on_conflict_do_nothing();
+        //~^ ERROR no method named `on_conflict_do_nothing` found
+    insert_into(users)
+        .values(&NewUser("Sean"))
+        .on_conflict_do_nothing()
+        .on_conflict_do_nothing();
+        //~^ ERROR no method named `on_conflict_do_nothing` found
+    insert_into(users)
+        .values(&NewUser("Sean")
+                .on_conflict(id, do_nothing())
+                .on_conflict_do_nothing());
+                //~^ ERROR no method named `on_conflict_do_nothing` found
+    insert_into(users)
+        .values(&NewUser("Sean").on_conflict(id, do_nothing()))
+        .on_conflict_do_nothing();
+        //~^ ERROR no method named `on_conflict_do_nothing` found
+    insert_into(users)
+        .values(&NewUser("Sean")
+                .on_conflict_do_nothing()
+                .on_conflict(id, do_nothing()));
+                //~^ ERROR no method named `on_conflict` found
+    insert_into(users)
+        .values(&NewUser("Sean")
+                .on_conflict(id, do_nothing())
+                .on_conflict(id, do_nothing()));
+                //~^ ERROR no method named `on_conflict` found
+    insert_into(users)
+        .values(&vec![NewUser("Sean").on_conflict_do_nothing()]);
+        //~^ ERROR UndecoratedInsertRecord
+    insert_into(users)
+        .values(&vec![&NewUser("Sean").on_conflict_do_nothing()]);
+        //~^ ERROR UndecoratedInsertRecord
+    insert_into(users)
+        .values(&vec![&NewUser("Sean").on_conflict(id, do_nothing())]);
+        //~^ ERROR UndecoratedInsertRecord
+    insert_into(users)
+        .values(&(name.eq("Sean").on_conflict_do_nothing(),));
+        //~^ ERROR UndecoratedInsertRecord
 }