+++ CHANGELOG.md
@@ -22,6 +22,9 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 * Added `replace_into(table).values(values)` as a replacement for
   `insert_or_replace(values).into(table)`.
 
+* Added `on_conflict_do_nothing` on `InsertStatement` as a replacement for
+  `on_conflict_do_nothing` on `Insertable` structs.
+
 ### Changed
 
 * The signatures of `QueryId`, `Column`, and `FromSqlRow` have all changed to
@@ -37,6 +40,9 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 * Deprecated `insert_or_replace(values).into(table)` in favor of
   `replace_into(table).values(values)`.
+
+* Deprecated `.values(x.on_conflict_do_nothing())` in favor of
+  `.values(x).on_conflict_do_nothing()`
 
 ### Removed